requirements.txt
streamlit
langchain
scikit-learn
pymupdf
pdfplumber
qdrant-client
tesserocr
pandas
numpy
openai
fastapi
uvicorn
langfuse
crewai
pypdf
sentence-transformers
langchain_community
faiss-cpu
matplotlib
seaborn
protobuf==3.20.3
langdetect
pymupdf
